question 2

+++ question3.h
@@ -1,11 +1,14 @@
 #ifndef QUESTION3_H
 #define QUESTION3_H
 
+#include <stdlib.h>
+#include <stdio.h>
+
 void testFunctionQ3();
 
-double Runs(unsigned *tab, int tabLenght, int valLenght);
+double Runs(unsigned long *tab, int tabLenght, int valLenght);
 
-double oneProportion(unsigned *tab, int tabLenght, int valLenght);
+double oneProportion(unsigned long *tab, int tabLenght, int valLenght);
 int numberOfContinousSuite(unsigned *tab, int tabLenght, int valLenght);
 
 #endif
